Restructure tests and test basket.subtotal
Now seems like a reasonable time to start testing against the actual
baskets mentioned in the assignment.

I don't think we any longer need an actual test for "add_product", as
it's so central to the whole operation, we can simply use it in our
setUp, which will obvious fail if that method fails.